pyproject.toml
[project]
name = "bs-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.5",
    "python-dotenv>=1.0.1",
    "supabase>=2.10.0",
    "uvicorn>=0.32.1",
]
